requirements.txt
python-dotenv
cloudinary
pymysql
qrcode
Pillow
